Rename path_policy.rs to safe_path.rs
Per code review feedback, the file `src/domain/path_policy.rs` was
renamed to `src/domain/safe_path.rs` to better match the struct name
and intent. This change updates the file and the corresponding
module imports in `src/domain/mod.rs`.
