+++ src/domain/mod.rs
@@ -1,8 +1,8 @@
 pub mod context_file;
 pub mod error;
-pub mod path_policy;
+pub mod safe_path;
 pub mod ports;
 pub mod snippet;
 
 pub use error::AppError;
-pub use path_policy::SafePath;
+pub use safe_path::SafePath;
