+++ src/domain/safe_path.rs
